requirements.txt
flask-bootstrap
flask-nav
flask-debug
flask-wtf
